// src/types/enums.ts
export enum FilterRangeMin {
  count = 1,
  year = 1999
}
export enum FilterRangeMax {
  count = 100,
  year = 2022
}
export enum FilterName {
  count = 'Count',
  year = 'Year',
  manufacturer = 'Manufacturer',
  color = 'Color',
  frets = 'Frets'
}
export enum FilterProperty {
  count = 'count',
  year = 'year',
  manufacturer = 'manufacturer',
  color = 'color',
  frets = 'fretsCount',
  isOnlyPopular = 'isOnlyPopular',
}